Choose text frame for writing place functionality
Gathers information of all text frames on a page in an open InDesign document, and then returns information about the largest one. This program is used by a program within BaseCode (Transfer, Place, and Export) that enables automatic placing of writing pieces in the magazine (by dragging and dropping a Word doc on the app's icon).